requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pytesseract>=0.3.8
Pillow>8.3.2
python-multipart>=0.0.5